Refactor poll into react component
